backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
pytest==7.4.3
httpx==0.25.2
trimesh
PyMuPDF==1.23.8
opencv-python==4.8.1.78
numpy==1.24.3
easyocr
matplotlib  # Optional: for debug plots (OD radius curve visualization)
openpyxl==3.1.5
pdfplumber==0.11.9  # PDF text extraction for engineering drawings
PyPDF2==3.0.1  # Fallback PDF parser
pytesseract  # OCR via Tesseract (simpler than EasyOCR, no PyTorch needed)
pillow  # Image processing for OCR
requests>=2.31.0
openai>=1.0.0
python-dotenv>=1.0.0
google-auth>=2.20.0
google-genai>=0.8.0